speech controller/requirement.txt
astroid==2.3.2
baidu-aip==2.2.18.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
future==0.18.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pbr==5.4.3
PyAudio==0.2.11
pylint==2.4.3
requests==2.22.0
six==1.12.0
SpeechRecognition==3.8.1
stevedore==1.31.0
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper-win==1.2.5
wrapt==1.11.2
